feat: handle evm wallet connection
