+++ packages/web/hooks/evm-wallet.ts
@@ -4,3 +4,5 @@ export const useConnectEvmWallet = useConnect;
 export const useEvmWalletAccount = useAccount;
 export const useSwitchEvmChain = useSwitchChain;
 export const useDisconnectEvmWallet = useDisconnect;
+
+export type ConnectEvmWalletReturn = ReturnType<typeof useConnectEvmWallet>;
